WriteOptions -> WriteModelOptions
Renamed it for better consistency with ReadModelOptions
