+++ CesiumGltfWriter/src/ImageWriter.h
@@ -2,13 +2,13 @@
 #include "JsonWriter.h"
 #include <CesiumGltf/Image.h>
 #include <CesiumGltf/WriteGLTFCallback.h>
-#include <CesiumGltf/WriteOptions.h>
+#include <CesiumGltf/WriteModelOptions.h>
 #include <vector>
 
 namespace CesiumGltf {
 void writeImage(
     const std::vector<Image>& images,
     CesiumGltf::JsonWriter& jsonWriter,
-    const WriteOptions& flags,
+    const WriteModelOptions& flags,
     WriteGLTFCallback writeGLTFCallback = noopGltfWriter);
 }
